+++ catroidTest/src/at/tugraz/ist/catroid/test/content/brick/PlaceAtBrickTest.java
@@ -29,16 +29,16 @@ public class PlaceAtBrickTest extends AndroidTestCase {
 
 	public void testNormalBehavior() {
 		Sprite sprite = new Sprite("testSprite");
-		assertEquals("Unexpected initial sprite x position", 0, sprite.costume.getXPosition());
-		assertEquals("Unexpected initial sprite y position", 0, sprite.costume.getYPosition());
+		assertEquals("Unexpected initial sprite x position", 0f, sprite.costume.getXPosition());
+		assertEquals("Unexpected initial sprite y position", 0f, sprite.costume.getYPosition());
 
 		PlaceAtBrick brick = new PlaceAtBrick(sprite, xPosition, yPosition);
 		brick.execute();
 
 		assertEquals("Incorrect sprite x position after PlaceAtBrick executed", xPosition,
-				sprite.costume.getXPosition());
+				(int) sprite.costume.getXPosition());
 		assertEquals("Incorrect sprite y position after PlaceAtBrick executed", yPosition,
-				sprite.costume.getYPosition());
+				(int) sprite.costume.getYPosition());
 	}
 
 	public void testNullSprite() {
@@ -58,16 +58,16 @@ public class PlaceAtBrickTest extends AndroidTestCase {
 		brick.execute();
 
 		assertEquals("PlaceAtBrick failed to place Sprite at maximum x integer value", Integer.MAX_VALUE,
-				sprite.costume.getXPosition());
+				(int) sprite.costume.getXPosition());
 		assertEquals("PlaceAtBrick failed to place Sprite at maximum y integer value", Integer.MAX_VALUE,
-				sprite.costume.getYPosition());
+				(int) sprite.costume.getYPosition());
 
 		brick = new PlaceAtBrick(sprite, Integer.MIN_VALUE, Integer.MIN_VALUE);
 		brick.execute();
 
 		assertEquals("PlaceAtBrick failed to place Sprite at minimum x integer value", Integer.MIN_VALUE,
-				sprite.costume.getXPosition());
+				(int) sprite.costume.getXPosition());
 		assertEquals("PlaceAtBrick failed to place Sprite at minimum y integer value", Integer.MIN_VALUE,
-				sprite.costume.getYPosition());
+				(int) sprite.costume.getYPosition());
 	}
 }
